✨ Refactor authentication decorators: move login_required to login_auth and clean up unused code
